+++ client/capi/base/header.h
@@ -3,7 +3,7 @@
 // Created Date: 18/05/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 21/05/2021
+// Last Modified: 14/12/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -15,4 +15,11 @@
 #define EXPORT_AUTD __declspec(dllexport)
 #else
 #define EXPORT_AUTD __attribute__((visibility("default")))
+#endif
+
+#ifdef __cplusplus
+#include <cstdint>
+#else
+#include <stdbool.h>
+#include <stdint.h>
 #endif