+++ client/capi/holo_gain/blas/c_api.cpp
@@ -3,7 +3,7 @@
 // Created Date: 11/09/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 11/09/2021
+// Last Modified: 12/12/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -14,6 +14,6 @@
 #include "autd3/gain/blas_backend.hpp"
 
 void AUTDBLASBackend(void** out) {
-  auto* b = BackendCreate(autd::gain::holo::BLASBackend::create());
+  auto* b = backend_create(autd::gain::holo::BLASBackend::create());
   *out = b;
 }