+++ client/examples/emulator.cpp
@@ -3,7 +3,7 @@
 // Created Date: 05/07/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 21/11/2021
+// Last Modified: 09/12/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -16,10 +16,10 @@
 
 int main() try {
   // Please see `dist/emulator/README.md`
-  auto autd = autd::Controller::create();
-  autd->geometry().add_device(autd::Vector3(0, 0, 0), autd::Vector3(0, 0, 0));
+  autd::Controller autd;
+  autd.geometry().add_device(autd::Vector3(0, 0, 0), autd::Vector3(0, 0, 0));
 
-  autd->open(autd::link::Emulator::create(50632, autd->geometry()));
+  autd.open(autd::link::Emulator::create(50632, autd.geometry()));
   return run(std::move(autd));
 } catch (std::exception& e) {
   std::cerr << e.what() << std::endl;