+++ client/lib/link/emulator/emulator.cpp
@@ -3,7 +3,7 @@
 // Created Date: 08/03/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 22/11/2021
+// Last Modified: 13/12/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -23,6 +23,7 @@
 
 #include "autd3/core/exception.hpp"
 #include "autd3/core/geometry.hpp"
+#include "autd3/core/interface.hpp"
 #include "autd3/core/link.hpp"
 
 namespace autd::link {
@@ -38,8 +39,8 @@ class EmulatorImpl final : public Emulator {
 
   void open() override;
   void close() override;
-  void send(const uint8_t* buf, size_t size) override;
-  void receive(uint8_t* rx, size_t buffer_len) override;
+  void send(const core::TxDatagram& tx) override;
+  void receive(core::RxDatagram& rx) override;
   bool is_open() override;
 
  private:
@@ -53,35 +54,31 @@ class EmulatorImpl final : public Emulator {
   sockaddr_in _addr = {};
 
   uint8_t _last_msg_id = 0;
-  std::vector<uint8_t> _geometry_buf;
+  core::TxDatagram _geometry_datagram;
 
-  static std::vector<uint8_t> init_geometry_buf(const core::Geometry& geometry) {
-    constexpr auto vec_size = 9 * sizeof(float);
-    const auto size = sizeof(core::GlobalHeader) + geometry.num_devices() * vec_size;
-    std::vector<uint8_t> buf;
-    buf.resize(size);
+  static core::TxDatagram init_geometry_datagram(const core::Geometry& geometry) {
+    core::TxDatagram buf(geometry.num_devices());
 
-    auto* const uh = reinterpret_cast<core::GlobalHeader*>(&buf[0]);
+    auto* const uh = buf.header();
     uh->msg_id = core::MSG_EMU_GEOMETRY_SET;
     uh->fpga_ctrl_flags = 0x00;
     uh->cpu_ctrl_flags = 0x00;
     uh->mod_size = 0x00;
 
-    auto* const cursor = reinterpret_cast<float*>(&buf[sizeof(core::GlobalHeader)]);
     for (const auto& device : geometry) {
-      const auto i = device.id();
+      auto* const cursor = reinterpret_cast<float*>(buf.body(device.id()));
       auto origin = device.begin()->position().cast<float>();
       auto right = device.x_direction().cast<float>();
       auto up = device.y_direction().cast<float>();
-      cursor[9 * i] = origin.x();
-      cursor[9 * i + 1] = origin.y();
-      cursor[9 * i + 2] = origin.z();
-      cursor[9 * i + 3] = right.x();
-      cursor[9 * i + 4] = right.y();
-      cursor[9 * i + 5] = right.z();
-      cursor[9 * i + 6] = up.x();
-      cursor[9 * i + 7] = up.y();
-      cursor[9 * i + 8] = up.z();
+      cursor[0] = origin.x();
+      cursor[1] = origin.y();
+      cursor[2] = origin.z();
+      cursor[3] = right.x();
+      cursor[4] = right.y();
+      cursor[5] = right.z();
+      cursor[6] = up.x();
+      cursor[7] = up.y();
+      cursor[8] = up.z();
     }
 
     return buf;
@@ -94,12 +91,13 @@ core::LinkPtr Emulator::create(const uint16_t port, const core::Geometry& geomet
 }
 
 EmulatorImpl::EmulatorImpl(const uint16_t port, const core::Geometry& geometry)
-    : _is_open(false), _port(port), _geometry_buf(init_geometry_buf(geometry)) {}
+    : _is_open(false), _port(port), _geometry_datagram(init_geometry_datagram(geometry)) {}
 
-void EmulatorImpl::send(const uint8_t* buf, const size_t size) {
-  const auto* header = reinterpret_cast<const core::GlobalHeader*>(buf);
+void EmulatorImpl::send(const core::TxDatagram& tx) {
+  const auto* header = reinterpret_cast<const core::GlobalHeader*>(tx.data());
   _last_msg_id = header->msg_id;
-  if (sendto(_socket, reinterpret_cast<const char*>(buf), static_cast<int>(size), 0, reinterpret_cast<sockaddr*>(&_addr), sizeof _addr) == -1)
+  if (sendto(_socket, reinterpret_cast<const char*>(tx.data()), static_cast<int>(tx.size()), 0, reinterpret_cast<sockaddr*>(&_addr), sizeof _addr) ==
+      -1)
     throw core::exception::LinkError("failed to send data");
 }
 
@@ -132,7 +130,7 @@ void EmulatorImpl::open() {
 #endif
 
   _is_open = true;
-  this->send(&this->_geometry_buf[0], this->_geometry_buf.size());
+  this->send(this->_geometry_datagram);
 }
 
 void EmulatorImpl::close() {
@@ -146,11 +144,11 @@ void EmulatorImpl::close() {
   _is_open = false;
 }
 
-void EmulatorImpl::receive(uint8_t* rx, size_t buffer_len) {
-  for (size_t i = 0; i < buffer_len; i += 2) rx[i + 1] = this->_last_msg_id;
+void EmulatorImpl::receive(core::RxDatagram& rx) {
+  for (auto& [_, msg_id] : rx) msg_id = this->_last_msg_id;
 
-  const auto set = [rx, buffer_len](const uint8_t value) {
-    for (size_t i = 0; i < buffer_len; i += 2) rx[i] = value;
+  const auto set = [&rx](const uint8_t value) {
+    for (auto& [ack, _] : rx) ack = value;
   };
 
   switch (this->_last_msg_id) {